+++ mayan/__init__.py
@@ -1,7 +1,7 @@
 __title__ = 'Mayan EDMS'
-__version__ = '4.6.2'
+__version__ = '4.7dev0'
 __build__ = 0x040602
-__build_string__ = 'v4.6.2_Mon Mar 4 14:08:58 2024 -0400'
+__build_string__ = 'v4.5.6-478-g347b5eaf3f4_Sat Mar 9 00:17:22 2024 -0400'
 __django_version__ = '4.2'
 __author__ = 'Roberto Rosario'
 __author_email__ = 'roberto.rosario@mayan-edms.com'