+++ mayan/apps/document_downloads/literals.py
@@ -1,5 +1,8 @@
 from django.utils.translation import gettext_lazy as _
 
+DEFAULT_DOCUMENT_FILE_DOWNLOAD_BACKEND = 'mayan.apps.document_downloads.download_backends.http.DownloadBackendDocumentFileFileDirectStorage'
+DEFAULT_DOCUMENT_FILE_DOWNLOAD_BACKEND_ARGUMENTS = None
+
 DEFAULT_DOCUMENT_FILE_DOWNLOAD_MESSAGE_BODY = _(
     'The document files have been compressed '
     'and are available for download using the '