+++ mayan/apps/document_downloads/settings.py
@@ -4,6 +4,8 @@ from mayan.apps.smart_settings.settings import setting_cluster
 
 from .classes import DocumentFileCompressor
 from .literals import (
+    DEFAULT_DOCUMENT_FILE_DOWNLOAD_BACKEND,
+    DEFAULT_DOCUMENT_FILE_DOWNLOAD_BACKEND_ARGUMENTS,
     DEFAULT_DOCUMENT_FILE_DOWNLOAD_MESSAGE_BODY,
     DEFAULT_DOCUMENT_FILE_DOWNLOAD_MESSAGE_SUBJECT
 )
@@ -12,6 +14,22 @@ setting_namespace = setting_cluster.do_namespace_add(
     label=_(message='Document downloads'), name='document_downloads'
 )
 
+setting_document_file_download_backend = setting_namespace.do_setting_add(
+    default=DEFAULT_DOCUMENT_FILE_DOWNLOAD_BACKEND,
+    global_name='DOCUMENT_DOWNLOADS_DOCUMENT_FILE_DOWNLOAD_BACKEND',
+    help_text=_(
+        'Path to the download subclass to use when downloading document '
+        'files.'
+    )
+)
+setting_document_file_download_backend_arguments = setting_namespace.do_setting_add(
+    default=DEFAULT_DOCUMENT_FILE_DOWNLOAD_BACKEND_ARGUMENTS,
+    global_name='DOCUMENT_DOWNLOADS_DOCUMENT_FILE_DOWNLOAD_BACKEND_ARGUMENTS',
+    help_text=_(
+        'Arguments to pass to the '
+        'DOCUMENT_DOWNLOADS_DOCUMENT_FILE_DOWNLOAD_BACKEND.'
+    )
+)
 setting_message_body_template = setting_namespace.do_setting_add(
     default=DEFAULT_DOCUMENT_FILE_DOWNLOAD_MESSAGE_BODY,
     global_name='DOCUMENT_FILE_DOWNLOAD_MESSAGE_BODY', help_text=_(