+++ mayan/apps/document_downloads/views.py
@@ -4,16 +4,20 @@ from mayan.apps.documents.models.document_file_models import DocumentFile
 from mayan.apps.documents.models.document_models import Document
 from mayan.apps.documents.views.document_file_views import DocumentFileListView
 from mayan.apps.organizations.utils import get_organization_installation_url
-from mayan.apps.views.generics import (
-    MultipleObjectFormActionView, SingleObjectDownloadView
-)
+from mayan.apps.storage.views.download_views import ViewSingleObjectBackendDownload
+from mayan.apps.views.generics import MultipleObjectFormActionView
 from mayan.apps.views.view_mixins import MultipleExternalObjectViewMixin
 
+from .events import event_document_file_downloaded
 from .forms import DocumentDownloadFormSet
 from .icons import (
     icon_document_download_multiple, icon_document_file_download_quick
 )
 from .permissions import permission_document_file_download
+from .settings import (
+    setting_document_file_download_backend,
+    setting_document_file_download_backend_arguments
+)
 from .tasks import task_document_file_compress
 
 
@@ -38,9 +42,9 @@ class DocumentDownloadView(
     success_message_singular = _(
         '%(count)d document file queued for download.'
     )
-    title_plural = _(message='Download files of %(count)d documents')
-    title_single = _(message='Download files of document: %(object)s')
-    title_singular = _(message='Download files of %(count)d document')
+    title_plural = _('Download files of %(count)d documents')
+    title_single = _('Download files of document: %(object)s')
+    title_singular = _('Download files of %(count)d document')
     view_icon = icon_document_download_multiple
 
     def get_extra_context(self):
@@ -117,20 +121,14 @@ class DocumentDownloadView(
         return super().view_action(form=form)
 
 
-class DocumentFileDownloadView(SingleObjectDownloadView):
+class DocumentFileDownloadView(ViewSingleObjectBackendDownload):
+    backend_arguments = setting_document_file_download_backend_arguments.value
+    backend_path = setting_document_file_download_backend.value
+    download_event_type = event_document_file_downloaded
     object_permission = permission_document_file_download
     pk_url_kwarg = 'document_file_id'
     source_queryset = DocumentFile.valid.all()
     view_icon = icon_document_file_download_quick
 
-    def get_download_file_object(self):
-        instance = self.get_object()
-        instance._event_action_object = instance.document
-        instance._event_actor = self.request.user
-        return instance.get_download_file_object()
-
-    def get_download_filename(self):
-        return self.object.filename
-
-    def get_download_mime_type_and_encoding(self, file_object):
-        return (self.object.mimetype, self.object.encoding)
+    def get_download_event_action_object(self):
+        return self.object.document