+++ mayan/apps/storage/dependencies.py
@@ -10,5 +10,8 @@ PythonDependency(
     module=__name__, name='extract-msg', version_string='==0.48.0'
 )
 PythonDependency(
+    module=__name__, name='google-cloud-storage', version_string='==2.10.0'
+)
+PythonDependency(
     module=__name__, name='pycryptodome', version_string='==3.20.0'
 )