+++ mayan/apps/storage/model_mixins.py
@@ -5,45 +5,40 @@ from django.db import models
 from django.template.defaultfilters import filesizeformat
 from django.utils.translation import gettext_lazy as _, gettext
 
-from mayan.apps.events.decorators import method_event
-from mayan.apps.events.event_managers import EventManagerMethodAfter
-
-from .events import event_download_file_downloaded
-
 logger = logging.getLogger(name=__name__)
 
 
-class DatabaseFileModelMixin(models.Model):
-    filename = models.CharField(
-        db_index=True, max_length=255, verbose_name=_(message='Filename')
-    )
-    datetime = models.DateTimeField(
-        auto_now_add=True, verbose_name=_(message='Date time')
-    )
-
-    class Meta:
-        abstract = True
-
-    def delete(self, *args, **kwargs):
-        name = self.file.name
-        self.file.close()
-
-        if name:
-            self.file.storage.delete(name=name)
-        return super().delete(*args, **kwargs)
-
+class ModelMixinFileFieldOpen:
     def open(self, **kwargs):
-        # Some storage class file do not provide a mode attribute.
+        # Some storage class do not provide a file mode attribute.
         # In that case default to read only in binary mode.
         # Python's default is read only in text format which does not work
         # for this use case.
         # https://docs.python.org/3/library/functions.html#open
-        mode = getattr(self.file.file, 'mode', 'rb')
 
-        open_kwargs = {
-            'mode': mode,
-            'name': self.file.name
-        }
+        field_file = getattr(self, 'file')
+
+        # Ensure the caller cannot specify an alternate filename.
+        name = kwargs.pop('name', None)
+
+        if name:
+            logger.critical(
+                'Caller specified an alternate filename "%s" for '
+                'object: %s. Verify that this is not an attempt '
+                'to access unauthorized files.', name, self
+            )
+
+        try:
+            getattr(field_file.file, 'mode', None)
+        except AttributeError:
+            # Storage does not support mode. Discard caller supplied mode.
+            kwargs.pop('mode', None)
+
+        name = field_file.name
+
+        open_kwargs = {'name': name}
+
+        open_kwargs.update(**kwargs)
 
         # Close the self.file object as Django generates a new descriptor
         # when the file field is accessed.
@@ -61,20 +56,33 @@ class DatabaseFileModelMixin(models.Model):
             >>> with open('/path/to/hello.world', 'r') as f:
             ...     instance.file = File(f)
         """
-        self.file.close()
-        self.file.file.close()
+        field_file.close()
+        field_file.file.close()
 
-        # Ensure the caller cannot specify an alternate filename.
-        name = kwargs.pop('name', None)
+        return self._open(**open_kwargs)
+
+
+class DatabaseFileModelMixin(ModelMixinFileFieldOpen, models.Model):
+    filename = models.CharField(
+        db_index=True, max_length=255, verbose_name=_(message='Filename')
+    )
+    datetime = models.DateTimeField(
+        auto_now_add=True, verbose_name=_(message='Date time')
+    )
+
+    class Meta:
+        abstract = True
+
+    def _open(self, **kwargs):
+        return self.file.storage.open(**kwargs)
+
+    def delete(self, *args, **kwargs):
+        name = self.file.name
+        self.file.close()
 
         if name:
-            logger.warning(
-                'Caller tried to specify an alternate filename: %s', name
-            )
-
-        open_kwargs.update(**kwargs)
-
-        return self.file.storage.open(**open_kwargs)
+            self.file.storage.delete(name=name)
+        return super().delete(*args, **kwargs)
 
     def save(self, *args, **kwargs):
         if not self.file:
@@ -89,14 +97,6 @@ class DatabaseFileModelMixin(models.Model):
 
 
 class DownloadFileBusinessLogicMixin:
-    @method_event(
-        event_manager_class=EventManagerMethodAfter,
-        event=event_download_file_downloaded,
-        target='self'
-    )
-    def get_download_file_object(self):
-        return self.open(mode='rb')
-
     def get_size_display(self):
         return filesizeformat(bytes_=self.file.size)